Add files via upload
We implemented the k-NN algorithm from scratch on python 2.7 to conduct the experiments for the project. k-NN is an instance-based, competitive learning, and lazy learning algorithm. Instance based algorithms, sometimes called memory-based learning, are those algorithms that, instead of performing explicit generalization, use the instances seen in the training as a comparison standard. For k-NN, the entire training dataset is the model. When a prediction is required for an unseen data instance, the k-NN algorithm will search through the training dataset for the k-most similar instances. k-NN is a competitive learning model because a majority vote is performed among the selected k records to determine the class label and then assigned it to the query record. k-NN is considered a lazy learning that does not build a model or function previously, but yields the closest k records of the training data set that have the highest similarity to the test (i.e., query record). The prediction attribute of the most similar instances is summarized and returned as the prediction for the unseen instance. The similarity measure is dependent on the type of data. For real-valued data, the Euclidean distance can be used. Other types of data such as categorical or binary data, Hamming distance can be used. In the case of regression problems, the average of the predicted attribute may be returned. In the case of classification, the most prevalent class may be returned.